agent/requirements.txt
anthropic>=0.57.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pypdf>=4.2.0
python-docx>=1.1.0
python-dotenv>=1.0.1
